Unplace signs only in current buffer
Commit 784835d introduced a bug when you got multiple buffers opened in
multiple windows.

If you unplace signs in all opened windows, they're going to be placed
back by `show_complexity()` only in current window. But all other
windows will jump for a few pixels to the left, which is ugly.

The proper fix for original bug (when even current buffer signs weren't
unplaced) would be to use `*` together with `file`:

```
:sign unplace * file={fname}
		Remove all placed signs in file {fname}.
```